合并 master 8-03 16:37 ~ 8-04 08:30 的 10 个提交（931d1b7..91107a7），
对应 master 发布 0.12.2：下线遍历测试 checkpoint 恢复能力 + daq-t1603 协议帧重构。

冲突解决策略：
- 配置类（VERSION/config.yml/wails.json/package.json/project.nsi）：取 master 0.12.2
- AGENTS.md/CLAUDE.md：GitNexus 索引统计取 master 最新同步
- traversal_checkpoint.go + 6 个 test 文件：取 master（b64a569 下线 checkpoint，
  与 lts/win7 方向一致；日志改用标准库 log/slog，符合"日志只用标准库"硬偏好）
- i18nStore.ts：取 master（删除 checkpoint 相关条目，统一 2 空格缩进）
- p1604_adapter.go：取 master 注释（对应 9d524ff 死连接判定口径统一）
- daq_t1603 协议帧（types.go/daq_t1603.go/daq_t1603_frame.go）：取 master
  （aff614b 重构 decodeTemperatures + int64 防溢出 + ApplyDaqT1603Config 双向对齐）

验证： stash 已保护本地 5 个未提交修改，merge 后将恢复。

+++ shared/device-sdk/go/daq/core/types.go
@@ -103,7 +103,7 @@ type DataPayload struct {
 	DeviceID          string    `json:"deviceId"`
 	Timestamp         int64     `json:"timestamp"`
 	HardwareTimestamp float64   `json:"hardwareTimestamp,omitempty"` // 设备硬件时间戳（秒.纳秒），0 表示无时间戳
-	SequenceNumber    int       `json:"sequenceNumber,omitempty"`    // 帧序号（HEAD=1 时每帧递增），0 表示未开启
+	SequenceNumber    int64     `json:"sequenceNumber,omitempty"`    // 帧序号（HEAD=1 时每帧递增，来源 uint32；int64 保证 32 位平台不溢出），0 表示未开启
 	Channels          []float64 `json:"channels"`
 	ChannelIndices    []int     `json:"channelIndices"`
 }
